+++ tracedemo/SDK/TKUIKit/ViewModel/TKTableViewRowM.h
@@ -11,7 +11,7 @@
 
 @interface TKTableViewRowM : NSObject
 
-@property (nonatomic,assign) NSInteger rowHeight;
+@property (nonatomic,assign) CGFloat rowHeight;
 @property (nonatomic,strong) NSObject * rowData;
 @property (nonatomic,assign) BOOL isEmptyData;
